test: relocate remaining Apple-engine unit tests to apple/core (#980)
Finishes the relocation started in #983, which moved 19 Apple-engine unit
tests into src/platforms/apple/core/__tests__/ but deferred two files that
compute runtime fs paths to native runner artifacts.

Moves the last two:
- recording-scripts.test.ts
- runner-client.test.ts

Both needed the relative import specifiers re-relativized for the deeper
location, plus one extra `../` on the runtime fs paths that resolve against
the repo root (the native ios-runner/ project and test/integration/support
still live at the repo root; #968 only moved the TypeScript engine):

- recording-scripts.test.ts: recordingScriptsDir / recordingTestSupportDir
  (`../../../../` -> `../../../../../`) -> ios-runner/.../RecordingScripts
  and test/integration/support.
- runner-client.test.ts: repoRoot (`../../../../` -> `../../../../../`),
  which feeds the project `.tmp` dir and the ios-runner xcodeproj path.

Behaviorless move. Verified: tsc --noEmit clean, both test files pass
(64 tests), layering check OK, oxfmt/oxlint clean, fallow audit clean.

closes #980
